types: don't use compiler types for {u,i}64
To make it less annoying for us to use it in logging.

+++ kernel/include/common/types.h
@@ -10,13 +10,13 @@
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int i32;
-typedef int64_t i64;
+typedef signed long long i64;
 
 // unsigned types
 typedef uint8_t u8;
 typedef uint16_t u16;
 typedef unsigned int u32;
-typedef uint64_t u64;
+typedef unsigned long long u64;
 
 typedef size_t ptr_t;
